Move SharedString into its own crate and use it in language_models_core
