+++ crates/agent_ui/src/agent_model_selector.rs
@@ -90,7 +90,7 @@ impl Render for AgentModelSelector {
         let model = self.selector.read(cx).delegate.active_model(cx);
         let model_name: SharedString = model
             .as_ref()
-            .map(|model| model.model.name().0.to_string().into())
+            .map(|model| model.model.name().0.clone())
             .unwrap_or_else(|| SharedString::from("Select a Model"));
 
         let provider_icon = model.as_ref().map(|model| model.provider.icon());
