+++ crates/ai_onboarding/src/agent_api_keys_onboarding.rs
@@ -34,7 +34,7 @@ impl ApiKeysWithProviders {
             .filter(|provider| {
                 provider.is_authenticated(cx) && provider.id() != ZED_CLOUD_PROVIDER_ID
             })
-            .map(|provider| (provider.icon(), provider.name().0.to_string().into()))
+            .map(|provider| (provider.icon(), provider.name().0.clone()))
             .collect()
     }
 }
