+++ crates/language_model_core/Cargo.toml
@@ -15,9 +15,8 @@ doctest = false
 [dependencies]
 anyhow.workspace = true
 cloud_llm_client.workspace = true
-derive_more.workspace = true
 futures.workspace = true
-gpui_util.workspace = true
+gpui_shared_string.workspace = true
 http_client.workspace = true
 partial-json-fixer.workspace = true
 schemars.workspace = true
