+++ crates/language_model_core/src/language_model_core.rs
@@ -2,7 +2,6 @@ mod provider;
 mod rate_limiter;
 mod request;
 mod role;
-mod shared_string;
 pub mod tool_schema;
 pub mod util;
 
@@ -25,9 +24,9 @@ pub use crate::provider::*;
 pub use crate::rate_limiter::*;
 pub use crate::request::*;
 pub use crate::role::*;
-pub use crate::shared_string::SharedString;
 pub use crate::tool_schema::LanguageModelToolSchemaFormat;
 pub use crate::util::{fix_streamed_json, parse_prompt_too_long, parse_tool_arguments};
+pub use gpui_shared_string::SharedString;
 
 #[derive(Clone, Debug)]
 pub struct LanguageModelCacheConfiguration {
